Module-1_Data_Integration/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
numpy==2.1.2
tqdm==4.66.5
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
praw==7.7.1
regex==2024.9.11
scikit-learn==1.5.2          # for later weeks (TF-IDF, dedupe)
spacy==3.7.5                 # for later weeks (NER)
nltk==3.9.1                  # for later weeks
streamlit